createservices

+++ pages/manage.js
@@ -8,32 +8,47 @@ import Userlist from '../src/components/Userlist';
 
 
 
-const Mange =(props)=>{
-    const {users ,createUser , updateUser, deleteUser } = props ;
-    const [selected , setselected ] = useState ();
-    const [formDAta,setFormData] = useState({}) ;
-    
+const Mange = (props) => {
+    const { users, createUser, updateUser, deleteUser, checkUser } = props;
+    const [selected, setselected] = useState();
+    const [formDAta, setFormData] = useState({});
 
-  
-    return( 
-        <div>
-            <Tapbar /> 
-            <Admin
-                handChange={(data) => setFormData(data)}
-                handCreate ={(data) => createUser(data)}
-                user = {users[selected]}    
-            />
-            <Userlist/>
-            {/* <User 
-                selected = {setselected}
-                users ={users}
-                handleClick = {(index)=> setselected(index)}
-                handleUpdate={(index) => updateUser(index,formDAta)}
-                handleDelete={(index)=> deleteUser(index)}
-            /> */}
-          
-            
+    if(checkUser === null){
+        return(
+            <div>login frist</div>
+        )
+    }
+    if (checkUser.role === 1) {
+
+
+        return (
+            <div>
+                <Tapbar />
+                <Admin
+                    handChange={(data) => setFormData(data)}
+                    handCreate={(data) => createUser(data)}
+                    user={users[selected]}
+                />
+
+                <User
+                    selected={setselected}
+                    users={users}
+                    handleClick={(index) => setselected(index)}
+                    handleUpdate={(index) => updateUser(index, formDAta)}
+                    handleDelete={(index) => deleteUser(index)}
+                />
+
+
+            </div>
+        )
+    }
+
+    else {
+        return (
+            <div>
+                you not Admin
         </div>
-    ) }
-    
-export default connect(state=>state.User, UserAction)(Mange);
+        )
+    }
+}
+export default connect(state => state.User, UserAction)(Mange);
